Major improvements in training the floating point VAE and VICreg, as well as training VAE on top of vicreg output. Added quantised VAE. WIP testing the quantised implementations.

+++ src/data/L1AD_datamodule.py
@@ -224,9 +224,9 @@ class L1ADDataModule(LightningDataModule):
         a bit faster.
         """
         dataset = L1ADDataset(
-            self.data_train,
+            self.data_train.float(),
             self.mask_train,
-            self.labels_train,
+            self.labels_train.float(),
             batch_size=self.batch_size_per_device,
             shuffler=self.shuffler,
         )
